fix: type to long

+++ src/main/kotlin/com/young/domain/item/dto/request/CreateItemRequest.kt
@@ -3,6 +3,6 @@ package com.young.domain.item.dto.request
 data class CreateItemRequest(
     val name: String,
     val description: String,
-    val price: Int,
-    val stock: Int
+    val price: Long,
+    val stock: Long
 )
